Handle errors

+++ app/src/main/java/com/mywork/topnews/model/ModelResponse.kt
@@ -6,15 +6,15 @@ import android.support.annotation.Nullable
 /**
  * Created by Pushpan on 3/08/18.
  */
-class ModelResponse(val status: Long, @Nullable val value: List<Article>, @Nullable val error: Throwable?) {
+class ModelResponse(val status: Int, @Nullable val value: List<Article>, @Nullable val error: Throwable?) {
 
     companion object {
         @JvmStatic
-        val Loading = 0L
+        val Loading = 0
         @JvmStatic
-        val Success = 1L
+        val Success = 1
         @JvmStatic
-        val Error = 2L
+        val Error = 2
 
         @JvmStatic
         fun success(value: List<Article>): ModelResponse = ModelResponse(Success, value,  null)
